Fix state/alarm deletion and lease invalidation bugs (#1)
This PR fixes several bugs in state and alarm management.

## Key Changes

### State Deletion Fixes
- **Fixed TTL comparison logic**: Changed `actor_state_expiration_time < now()` to `> now()` in both PostgreSQL and SQLite `DeleteState` operations. The previous logic would only delete expired state, not live state with active TTLs.
- **Added test coverage**: New test case verifies that `DeleteState` correctly removes state with a live TTL and returns `ErrNoState` on subsequent operations.

### Alarm Lease Invalidation
- **Fixed stale lease handling in `SetAlarm`**: When replacing an existing alarm, the old alarm ID mapping is now deleted from `AlarmsByID` to prevent stale leases from resolving to phantom alarms.
- **Added comprehensive test**: New test case "replacing an alarm invalidates a previously-issued lease" verifies that:
  - Stale leases cannot be used with `GetLeasedAlarm`, `UpdateLeasedAlarm`, or `DeleteLeasedAlarm`
  - The replacement alarm exists with correct properties
  - This prevents regressions where stale leases could read/update/delete phantom alarms

### Alarm Lease Renewal Fixes
- **Fixed empty host list handling**: Added early return in SQLite's `RenewAlarmLeases` when no hosts are provided, preventing invalid SQL with empty `IN ()` clauses.
- **Added test coverage**: New test verifies that renewal with an empty host list returns no leases without error.

### PostgreSQL Migration Improvements
- **Simplified advisory lock handling**: Changed from session-level `pg_try_advisory_lock` to transaction-level `pg_try_advisory_xact_lock` in the fetch-and-lease-alarms function. This automatically releases locks when the transaction ends, preventing lock leaks on pooled connections that could make actors permanently un-allocatable.
- **Removed manual lock release code**: Eliminated exception handling and explicit `pg_advisory_unlock` calls since transaction-level locks are automatically released.

### Code Quality Improvements
- **Fixed migration array sizing**: Corrected `len(entries)` to `len(names)` in both PostgreSQL and SQLite migration setup (entries is unused after sorting).
- **Improved error handling**: Changed panic to proper error return in `RenewAlarmLeases` when lease ID type assertion fails.
- **Updated imports**: Replaced `database/sql` with `github.com/jackc/pgx/v5` for proper error type checking (`pgx.ErrNoRows` instead of `sql.ErrNoRows`).

+++ host/messaging_test.go
@@ -477,7 +477,7 @@ func TestLockAndInvokeFn(t *testing.T) {
 		}
 
 		// Verify all actors were created
-		assert.Equal(t, numActors, int(host.actors.Len()))
+		assert.Equal(t, uintptr(numActors), host.actors.Len())
 	})
 
 	t.Run("concurrent invocations with same actor are serialized", func(t *testing.T) {
@@ -528,7 +528,7 @@ func TestLockAndInvokeFn(t *testing.T) {
 		assert.Len(t, executionOrder, numInvocations)
 
 		// Only one actor should exist
-		assert.Equal(t, 1, int(host.actors.Len()))
+		assert.Equal(t, uintptr(1), host.actors.Len())
 	})
 
 	t.Run("parent context cancellation", func(t *testing.T) {
